- Summary: Fix Bugzilla Bug 152678.ParameterDialog cannot be build if compiler level is 5.0
- Bugzilla Bug (s) Resolved: 152678

- Description: Change the code to compatible with JDK 5.0

- Tests Description  : Manual Test

- Files Edited:
@changelist
- Files Added:
@addlist
- Files Deleted:
@deletelist
- Notes to Build Team:

- Notes to Developers:

- Notes to QA:

- Notes to Documentation:

+++ UI/org.eclipse.birt.report.designer.ui/src/org/eclipse/birt/report/designer/ui/dialogs/ParameterDialog.java
@@ -1929,15 +1929,15 @@ public class ParameterDialog extends BaseDialog
 		}
 		if ( v1 instanceof Double && v2 instanceof Double )
 		{
-			return ( (Double) v1 ).compareTo( v2 ) == 0;
+			return ( (Double) v1 ).compareTo( (Double) v2 ) == 0;
 		}
 		if ( v1 instanceof BigDecimal && v2 instanceof BigDecimal )
 		{
-			return ( (BigDecimal) v1 ).compareTo( v2 ) == 0;
+			return ( (BigDecimal) v1 ).compareTo( (BigDecimal) v2 ) == 0;
 		}
 		if ( v1 instanceof Integer && v2 instanceof Integer )
 		{
-			return ( (Integer) v1 ).compareTo( v2 ) == 0;
+			return ( (Integer) v1 ).compareTo( (Integer) v2 ) == 0;
 		}
 		return v1.equals( v2 );
 	}
